pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Deep Research Agent - An autonomous research assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.10.0",
    "html2text>=2025.4.15",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "trafilatura>=2.0.0",
    "streamlit>=1.42.0",
]

[project.scripts]
deep-research = "deep_research.cli:main"

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]
